# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Weather Service","description":"Service for weather information.","version":"1.0"},"servers":[{"url":"https://dev.myweatherdomain.com","description":"Weather service"},{"url":"http://localhost:8080","description":"Weather service"}],"paths":{"/liveness":{"get":{"tags":["healthcheck"],"summary":"liveness healthcheck","operationId":"healthcheck#liveness","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LivenessResponseBody"},"example":{"status":"up"}}}}}}},"/readiness":{"get":{"tags":["healthcheck"],"summary":"readiness healthcheck","operationId":"healthcheck#readiness","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LivenessResponseBody"},"example":{"status":"up"}}}},"503":{"description":"Service Unavailable response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LivenessResponseBody"},"example":{"status":"status"}}}}}}},"/v1/weather":{"post":{"tags":["weather"],"summary":"WeatherQuery weather","operationId":"weather#WeatherQuery","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherQueryRequestBody"},"example":{"lat":51.50853,"lon":-0.12574,"units":"metric"}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherQueryResult"},"example":{"feels_like":0.28940177100794445,"pressure":4182469846481591513,"temp":0.9800462962300538,"wind_speed":4837219454189177381}}}},"400":{"description":"Bad Request response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherQueryResult"},"example":{"feels_like":0.4918492209488689,"pressure":3456525917261053504,"temp":0.7073857363576652,"wind_speed":5497966619188418024}}}},"500":{"description":"Internal Server Error response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherQueryResult"},"example":{"feels_like":0.8254494184927246,"pressure":1265099336773306726,"temp":0.5314365495128938,"wind_speed":374781901373090807}}}}}}}},"components":{"schemas":{"LivenessResponseBody":{"type":"object","properties":{"status":{"type":"string","example":"up"}},"example":{"status":"up"},"required":["status"]},"WeatherQueryRequestBody":{"type":"object","properties":{"lat":{"type":"number","description":"Latitude coordinate of location.","example":51.50853,"format":"double"},"lon":{"type":"number","description":"Longitude coordinate of location.","example":-0.12574,"format":"double"},"units":{"type":"string","description":"Units of measurement. Possible values are 'standard', 'metric' or 'imperial'.","example":"metric"}},"example":{"lat":51.50853,"lon":-0.12574,"units":"metric"},"required":["lat","lon"]},"WeatherQueryResult":{"type":"object","properties":{"feels_like":{"type":"number","example":0.39810404641170455,"format":"double"},"pressure":{"type":"integer","example":7404522694356190867,"format":"int64"},"temp":{"type":"number","example":0.5375015712497628,"format":"double"},"wind_speed":{"type":"integer","example":6714187406412673715,"format":"int64"}},"example":{"feels_like":0.6274746697076189,"pressure":6040948167692002309,"temp":0.3837911121532482,"wind_speed":2183901827372585546},"required":["temp"]}}},"tags":[{"name":"healthcheck","description":"The healthcheck service is used report on the liveness and readiness status of the weather service."},{"name":"weather","description":"The weather service returns information for a given location."}]}